Rename coco and images_dir
