+++ datumaro/datumaro/components/extractors/__init__.py
@@ -5,7 +5,7 @@
 
 from datumaro.components.extractors.datumaro import DatumaroExtractor
 
-from datumaro.components.extractors.ms_coco import (
+from datumaro.components.extractors.coco import (
     CocoImageInfoExtractor,
     CocoCaptionsExtractor,
     CocoInstancesExtractor,
@@ -29,7 +29,7 @@ from datumaro.components.extractors.voc import (
 from datumaro.components.extractors.yolo import YoloExtractor
 from datumaro.components.extractors.tfrecord import DetectionApiExtractor
 from datumaro.components.extractors.cvat import CvatExtractor
-from datumaro.components.extractors.images_dir import ImagesDirExtractor
+from datumaro.components.extractors.image_dir import ImageDirExtractor
 
 items = [
     ('datumaro', DatumaroExtractor),
@@ -58,5 +58,5 @@ items = [
 
     ('cvat', CvatExtractor),
 
-    ('images_dir', ImagesDirExtractor),
+    ('image_dir', ImageDirExtractor),
 ]
