+++ datumaro/datumaro/components/formats/coco.py
