+++ datumaro/datumaro/components/importers/__init__.py
@@ -4,23 +4,18 @@
 # SPDX-License-Identifier: MIT
 
 from datumaro.components.importers.datumaro import DatumaroImporter
-from datumaro.components.importers.ms_coco import CocoImporter
-
-from datumaro.components.importers.voc import (
-    VocImporter,
-    VocResultsImporter,
-)
-
+from datumaro.components.importers.coco import CocoImporter
+from datumaro.components.importers.voc import VocImporter, VocResultsImporter
 from datumaro.components.importers.tfrecord import DetectionApiImporter
 from datumaro.components.importers.yolo import YoloImporter
 from datumaro.components.importers.cvat import CvatImporter
-from datumaro.components.importers.images_dir import ImagesDirImporter
+from datumaro.components.importers.image_dir import ImageDirImporter
 
 
 items = [
     ('datumaro', DatumaroImporter),
 
-    ('ms_coco', CocoImporter),
+    ('coco', CocoImporter),
 
     ('voc', VocImporter),
     ('voc_results', VocResultsImporter),
@@ -31,5 +26,5 @@ items = [
 
     ('cvat', CvatImporter),
 
-    ('images_dir', ImagesDirImporter),
+    ('image_dir', ImageDirImporter),
 ]
